requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
bootstrap-discord-bot==1.0.8
certifi==2020.6.20
chardet==3.0.4
discord==1.0.1
discord.py==1.5.0
gitdb==4.0.5
GitPython==3.1.9
idna==2.10
multidict==4.7.6
nbtlib==1.8.1
numpy==1.19.2
psutil==5.7.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
smmap==3.0.4
urllib3==1.25.10
websockets==8.1
yarl==1.6.0
